Removed status checking Job and minor refactor
